% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_1.R
\name{ex_1}
\alias{ex_1}
\title{Function created for problem 1}
\usage{
ex_1()
}
\arguments{
\item{No}{parameters needed}
}
\value{
One letter
}
\description{
This function returns a randomly chosen letter of the alphabet
}
